requirements.txt
asgiref==3.8.1
boto3==1.34.145
botocore==1.34.145
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.7
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-js-asset==2.2.0
django-s3-upload==1.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
jmespath==1.0.1
pillow==10.4.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0        
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
